+++ src/lib.rs
@@ -31,7 +31,7 @@
 //!
 //!     // let driver = NetworkDriver::open("192.168.1.248", 9100)?;
 //!     let driver = ConsoleDriver::open(true);
-//!     Printer::new(driver, Protocol::default())
+//!     Printer::new(driver, Protocol::default(), None)
 //!         .debug_mode(Some(DebugMode::Dec))
 //!         .init()?
 //!         .smoothing(true)?
@@ -66,7 +66,7 @@
 //!
 //!     // let driver = NetworkDriver::open("192.168.1.248", 9100)?;
 //!     let driver = ConsoleDriver::open(true);
-//!     Printer::new(driver, Protocol::default())
+//!     Printer::new(driver, Protocol::default(), None)
 //!         .debug_mode(Some(DebugMode::Hex))
 //!         .init()?
 //!         .ean13_option(
@@ -97,7 +97,7 @@
 //!
 //!     // let driver = NetworkDriver::open("192.168.1.248", 9100)?;
 //!     let driver = ConsoleDriver::open(true);
-//!     Printer::new(driver, Protocol::default())
+//!     Printer::new(driver, Protocol::default(), None)
 //!         .debug_mode(Some(DebugMode::Hex))
 //!         .init()?
 //!         .qrcode_option(
@@ -123,7 +123,7 @@
 //!
 //!     // let driver = NetworkDriver::open("192.168.1.248", 9100)?;
 //!     let driver = ConsoleDriver::open(true);
-//!     let mut printer = Printer::new(driver, Protocol::default());
+//!     let mut printer = Printer::new(driver, Protocol::default(), None);
 //!     printer.debug_mode(Some(DebugMode::Hex))
 //!         .init()?
 //!         .bit_image_option(